extract logic to webpack plugin
